+++ packages/BoEmoji/index.js
@@ -1,0 +1,7 @@
+import BoEmoji from './src/main';
+
+BoEmoji.install = function(Vue) {
+  Vue.component(BoEmoji.name, BoEmoji)
+};
+
+export default BoEmoji;