+++ packages/BoImageUpload/index.js
@@ -1,0 +1,7 @@
+import BoImageUpload from './src/main';
+
+BoImageUpload.install = function(Vue) {
+  Vue.component(BoImageUpload.name, BoImageUpload)
+};
+
+export default BoImageUpload;